refactor unit test and impl test shell script
