+++ src/kstd/status.rs
@@ -16,8 +16,8 @@ use std::fmt;
 
 #[derive(Debug)]
 pub struct Status {
-    code: Code,
-    message: String,
+    pub(crate) code: Code,
+    pub(crate) message: String,
 }
 
 /// TODO: remove allow dead code.
@@ -58,56 +58,5 @@ impl Status {
 impl fmt::Display for Status {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         write!(f, "{:?}: {}", self.code, self.message)
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_status_ok() {
-        let status = Status::ok();
-        assert_eq!(status.code, Code::Ok);
-        assert_eq!(status.message, "");
-    }
-
-    #[test]
-    fn test_status_timeout() {
-        let status = Status::timeout("Connection timeout");
-        assert_eq!(status.code, Code::Timeout);
-        assert_eq!(status.message, "Connection timeout");
-    }
-
-    #[test]
-    fn test_empty_timeout_message() {
-        let status = Status::timeout("");
-        assert_eq!(status.code, Code::Timeout);
-        assert_eq!(status.message, "");
-    }
-
-    #[test]
-    fn test_to_string() {
-        let status = Status::timeout("Operation timeout");
-        assert_eq!(status.to_string(), "Timeout: Operation timeout");
-    }
-
-    #[test]
-    fn test_display_trait() {
-        let status = Status::ok();
-        assert_eq!(format!("{}", status), "Ok: ");
-
-        let status = Status::timeout("Request timeout");
-        assert_eq!(format!("{}", status), "Timeout: Request timeout");
-    }
-
-    #[test]
-    fn test_debug_trait() {
-        let status = Status::ok();
-        assert!(format!("{:?}", status).contains("Ok"));
-
-        let status = Status::timeout("Network timeout");
-        assert!(format!("{:?}", status).contains("Timeout"));
-        assert!(format!("{:?}", status).contains("Network timeout"));
     }
 }
