+++ src/kstd/tests/mod.rs
@@ -1,0 +1,3 @@
+mod lock_mgr_test;
+mod slice_test;
+mod status_test;
