+++ src/storage/tests/mod.rs
@@ -1,0 +1,8 @@
+mod base_data_value_format_test;
+mod base_key_format_test;
+mod lru_cache_test;
+mod slot_indexer_test;
+mod storage_define_test;
+mod storage_murmur3_test;
+mod util_test;
+// mod column_family;
